firstly created Methods to use in Driver Controlled
Created HardwareMaps for both DriveTrains yet only speaking to the 4 elementary Motors. Aswell created a Tools class yet only used to speak to all 4 Moters with one Command. Also created Driver Controlled Class, with the (YET UNTESTED ASWELL AS THE OTHER CLASSES) to gain the ability to drive in 4 directions.